remove misplaced cast
git-svn-id: http://svn.ratbox.org/svnroot/trunk@373 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ src/channel.c
@@ -2933,7 +2933,7 @@ void send_user_joins(struct Client *cptr, struct Client *user)
         if (!match(++mask, cptr->name))
           continue;
       clen = strlen(chptr->chname);
-      if (clen > (size_t) BUFSIZE - 7 - len)
+      if (clen > BUFSIZE - 7 - len)
         {
           if (cnt)
             sendto_one(cptr, "%s", buf);
